+++ project/product_delete.php
@@ -1,26 +1,34 @@
 <?php
 // include database connection
 include 'config/database.php';
-try {     
+try {
     // get record ID
     // isset() is a PHP function used to verify if a value is there or not
-    $id=isset($_GET['id']) ? $_GET['id'] :  die('ERROR: Record ID not found.');
+    $id = isset($_GET['id']) ? $_GET['id'] :  die('ERROR: Record ID not found.');
 
     // delete query
-    $query = "DELETE FROM products WHERE id = ?";
-    $stmt = $con->prepare($query);
-    $stmt->bindParam(1, $id);
-     
-    if($stmt->execute()){
-        // redirect to read records page and
-        // tell the user record was deleted
-        header('Location: product_read.php?action=deleted');
-    }else{
-        die('Unable to delete record.');
+    $query_order_list = "SELECT * FROM order_detail WHERE product_id = ?";
+    $stmt_order_list = $con->prepare($query_order_list);
+    $stmt_order_list->bindParam(1, $id);
+    $stmt_order_list->execute();
+    $num = $stmt_order_list->rowCount();
+    if ($num > 0) {
+        header('Location: customer_read.php?action=cancel');
+    } else {
+        $query = "DELETE FROM products WHERE id = ?";
+        $stmt = $con->prepare($query);
+        $stmt->bindParam(1, $id);
+
+        if ($stmt->execute()) {
+            // redirect to read records page and
+            // tell the user record was deleted
+            header('Location: product_read.php?action=deleted');
+        } else {
+            die('Unable to delete record.');
+        }
     }
 }
 // show error
-catch(PDOException $exception){
+catch (PDOException $exception) {
     die('ERROR: ' . $exception->getMessage());
 }
-?>
